backend/pyproject.toml
[project]
name = "turing-qna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastembed>=0.7.3",
    "gradio>=5.49.1",
    "pydantic-ai>=1.11.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
]
